requirements.txt
streamlit
requests
beautifulsoup4
transformers
fpdf
deep-translator
python-docx
qrcode
